+++ tests/test_layout.py
@@ -1,4 +1,3 @@
-import django
 import pytest
 from django import forms
 from django.forms.models import formset_factory, modelformset_factory
@@ -243,10 +242,7 @@ def test_formset_layout():
     )
 
     html = render_crispy_form(form=formset, helper=helper, context={"csrf_token": "aTestToken"})
-    if django.VERSION < (5, 0):
-        result = "test_formset_layout lt50.html"
-    else:
-        result = "test_formset_layout.html"
+    result = "test_formset_layout.html"
     assert parse_expected(result) == parse_html(html)